requirements.txt
certifi==2020.6.20
chardet==3.0.4
fedex==2.4.1
future==0.18.2
idna==2.10
numpy==1.19.1
pandas==1.1.1
probableparsing==0.0.1
PyQt5==5.15.0
PyQt5-sip==12.8.1
pyserial==3.4
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
suds-jurko==0.6
urllib3==1.25.10
usaddress==0.5.10
